// hypertrace-ui/projects/components/src/table/cells/types/core-table-cell-renderer-type.ts
export const enum CoreTableCellRendererType {
  Checkbox = 'checkbox',
  Code = 'code',
  Duration = 'duration',
  Icon = 'icon',
  Number = 'number',
  RowExpander = 'row-expander',
  RelativeTimestamp = 'relative-timestamp',
  StringArray = 'string-array',
  StringEnum = 'string-enum',
  Text = 'text',
  TextWithCopyAction = 'text-with-copy-action',
  Timestamp = 'timestamp',
  TimeAgo = 'time-ago',
}